requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
joblib==1.5.1
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
numba==0.61.2
numpy==2.2.6
optuna==4.5.0
packaging==25.0
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
shap==0.48.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.29.0
watchfiles==1.1.0
websockets==15.0.1
xgboost==3.0.4
